--- a/perf/flame.jl
+++ b/perf/flame.jl
@@ -1,12 +1,14 @@
-example_dir = joinpath(dirname(@__DIR__), "examples")
-
-import Profile
+ca_dir = joinpath(dirname(@__DIR__));
+include(joinpath(ca_dir, "examples", "hybrid", "cli_options.jl"));
 
 ENV["CI_PERF_SKIP_RUN"] = true # we only need haskey(ENV, "CI_PERF_SKIP_RUN") == true
 
-filename = joinpath(example_dir, "hybrid", "driver.jl")
+filename = joinpath(ca_dir, "examples", "hybrid", "driver.jl")
+dict = parsed_args_per_job_id(; trigger = "benchmark.jl")
+parsed_args = dict["perf_target_unthreaded"];
 
-try
+
+try # capture integrator
     include(filename)
 catch err
     if err.error !== :exit_profile
@@ -15,9 +17,10 @@ catch err
 end
 
 OrdinaryDiffEq.step!(integrator) # compile first
+import Profile
 Profile.clear_malloc_data()
 prof = Profile.@profile begin
-    for _ in 1:5
+    for _ in 1:10
         OrdinaryDiffEq.step!(integrator)
     end
 end