This PR

 - Addresses @cmbengue's request to polish / automate a performance analysis report for single processor (unthreaded + multithreaded) for our nearest neighbor to amip. At the moment, the performance targets are missing edmf-- which I'll try to fix in a subsequent PR, after which the performance report will auto-update.

 - Adds/uses some benchmark utilities developed in Thermodynamics
 - Adds two benchmarks: unthreaded (adjusted existing benchmark script) + multithreaded (new script)

For now, we're explicitly avoiding IO, however, we should revisit this soon. Also, our benchmark for `step!` is not "robust" to callbacks-- we may call some of the callbacks depending on the interval / frequency specified. This should also be revisited.

Here's an example output (for the unthreaded):

```
Benchmarking Wfact...
Benchmarking linsolve...
Benchmarking implicit_tendency!...
Benchmarking remaining_tendency!...
Benchmarking step!...
┌─────────────────────┬────────────┬──────────┬────────────┬────────────┬────────────┬────────────┬───────────┐
│ Function            │     Memory │   allocs │       Time │       Time │       Time │       Time │ N-samples │
│                     │   estimate │ estimate │        min │        max │       mean │     median │           │
├─────────────────────┼────────────┼──────────┼────────────┼────────────┼────────────┼────────────┼───────────┤
│ Wfact               │ 335.53 MiB │   259207 │ 528.069 ms │ 731.836 ms │ 605.385 ms │ 592.255 ms │         8 │
│ linsolve            │ 205.67 MiB │   280806 │ 277.099 ms │ 405.817 ms │ 316.950 ms │ 287.764 ms │        10 │
│ implicit_tendency!  │  13.67 KiB │       51 │ 649.578 ms │ 674.674 ms │ 658.891 ms │ 657.224 ms │         8 │
│ remaining_tendency! │  13.82 MiB │    43442 │    2.690 s │    2.697 s │    2.693 s │    2.693 s │         2 │
│ step!               │ 838.18 MiB │   951620 │   13.180 s │   13.180 s │   13.180 s │   13.180 s │         1 │
└─────────────────────┴────────────┴──────────┴────────────┴────────────┴────────────┴────────────┴───────────┘
```

Note that some of these functions call other functions, so there's a scope understanding that's required for interpreting results. `step!` is the outer-most call and can be considered the wall time per timestep (note that a single `step!` performs multiple `wfact`, `linsolve`, `implicit_tendency!` and `remaining_tendency!` calls).

These results will be available on every merged PR in the "performance targets" jobs (threaded / unthreaded).

I'm using BenchmarkTools to perform the benchmarks, we could call multiple times to force a larger sample size, but the existing config will automatically increase the sample size as we improve performance.