@charleskawczynski 2022-07-14
Looks like we need some sort of config adjustment to get accurate numbers for the threading case. We can follow up on that and get the report in now.